refactor dpm campaign definition lookups
